+++ src/common/function/src/aggrs/aggr_wrapper.rs
@@ -21,14 +21,17 @@
 //! `foo_merge`'s input args is the same as `foo_state`'s, and its output is the same as `foo`'s.
 //!
 
-use std::collections::BTreeMap;
 use std::sync::Arc;
 
 use arrow::array::StructArray;
 use datafusion_common::ScalarValue;
 use datafusion_expr::function::StateFieldsArgs;
-use datafusion_expr::{Accumulator, AggregateUDF, AggregateUDFImpl, Signature, TypeSignature};
+use datafusion_expr::{Accumulator, AggregateUDF, AggregateUDFImpl, Signature};
 use datatypes::arrow::datatypes::{DataType, Field};
+
+use crate::aggrs::aggr_wrapper::type_lookup::StateTypeLookup;
+
+mod type_lookup;
 
 /// Returns the name of the state function for the given aggregate function name.
 /// The state function is used to compute the state of the aggregate function.
@@ -250,287 +253,6 @@ impl Accumulator for StateAccum {
     }
 }
 
-/// Find a valid input types from given state types.
-///
-/// Useful for merge functions that need to merge to use correct accumulator to merge states.
-#[derive(Debug, Default, Clone, PartialEq, Eq)]
-pub struct StateTypeLookup {
-    /// A mapping from state types to input types.
-    pub mapping: BTreeMap<Vec<DataType>, Vec<DataType>>,
-}
-
-impl StateTypeLookup {
-    /// Returns the input types for the given state types.
-    ///
-    /// Also handle certain data types that have extra args, like `Decimal128/256`, assuming only one of
-    /// the `Decimal128/256` in the state types, and replace it with the same type in the returned input types.
-    pub fn get_input_types(
-        &self,
-        state_types: &[DataType],
-    ) -> datafusion_common::Result<Option<Vec<DataType>>> {
-        let state_types = if state_types.len() == 1 && matches!(state_types[0], DataType::Struct(_))
-        {
-            // unpack single struct type
-            let struct_type = state_types[0].clone();
-            if let DataType::Struct(struct_ty) = struct_type {
-                let state_types = struct_ty
-                    .iter()
-                    .map(|f| f.data_type().clone())
-                    .collect::<Vec<_>>();
-                state_types
-            } else {
-                state_types.to_vec()
-            }
-        } else {
-            state_types.to_vec()
-        };
-
-        let is_decimal =
-            |ty: &DataType| matches!(ty, DataType::Decimal128(_, _) | DataType::Decimal256(_, _));
-        let need_change_ty = state_types.iter().any(is_decimal);
-        if !need_change_ty {
-            return Ok(self.mapping.get(&state_types).cloned());
-        }
-
-        for (k, v) in self.mapping.iter() {
-            if k.len() != state_types.len() {
-                continue;
-            }
-            let mut is_equal_types = true;
-            for (k, s) in k.iter().zip(state_types.iter()) {
-                let is_equal = match (k, s) {
-                    (DataType::Decimal128(_, _), DataType::Decimal128(_, _)) => true,
-                    (DataType::Decimal256(_, _), DataType::Decimal256(_, _)) => true,
-                    (a, b) => a == b,
-                };
-                if !is_equal {
-                    is_equal_types = false;
-                    break;
-                }
-            }
-
-            if is_equal_types {
-                // replace `Decimal` in input_types with same `Decimal` type in `state_types`
-                // Expect only one of the `Decimal128/256` in the state_types.
-                // If there are multiple, this logic might need to be revisited.
-                let mut input_types = v.clone();
-                let decimal_types_in_state: Vec<&DataType> =
-                    state_types.iter().filter(|ty| is_decimal(ty)).collect();
-
-                if decimal_types_in_state.len() > 1 {
-                    return Err(datafusion_common::DataFusionError::Internal(format!(
-                        "Multiple Decimal types found in state types, expected at most one: {:?}",
-                        state_types
-                    )));
-                }
-
-                if let Some(new_decimal_type) = decimal_types_in_state.first() {
-                    input_types = input_types
-                        .into_iter()
-                        .map(|ty| {
-                            if is_decimal(&ty) {
-                                (*new_decimal_type).clone()
-                            } else {
-                                ty
-                            }
-                        })
-                        .collect::<Vec<_>>();
-                }
-                return Ok(Some(input_types));
-            }
-        }
-        Ok(None)
-    }
-}
-
-/// TODO(discord9): put this into a separate file or register side by side with  `AggregateUDF` impl directly
-#[allow(unused)]
-fn get_possible_state_to_input_types(
-    inner: &AggregateUDF,
-) -> datafusion_common::Result<StateTypeLookup> {
-    let mut mapping = BTreeMap::new();
-
-    fn update_input_types(
-        inner: &AggregateUDF,
-        mapping: &mut BTreeMap<Vec<DataType>, Vec<DataType>>,
-        input_types: &[DataType],
-    ) -> datafusion_common::Result<()> {
-        let ret_type = inner.return_type(input_types)?;
-        let state_fields_args = StateFieldsArgs {
-            name: inner.name(),
-            input_types,
-            return_type: &ret_type,
-            ordering_fields: &[],
-            is_distinct: false,
-        };
-        let state_fields = inner.state_fields(state_fields_args)?;
-        mapping.insert(
-            state_fields.iter().map(|f| f.data_type().clone()).collect(),
-            input_types.to_vec(),
-        );
-        Ok(())
-    }
-
-    fn parse_ty_sig(
-        inner: &AggregateUDF,
-        sig: &TypeSignature,
-        mapping: &mut BTreeMap<Vec<DataType>, Vec<DataType>>,
-    ) -> datafusion_common::Result<()> {
-        match &sig {
-            TypeSignature::Exact(input_types) => {
-                let _ = update_input_types(inner, mapping, input_types);
-            }
-            TypeSignature::OneOf(type_sig) => {
-                for sig in type_sig {
-                    parse_ty_sig(inner, sig, mapping)?;
-                }
-            }
-            TypeSignature::Numeric(len) => {
-                for ty in DataTypeIterator::all_flat().filter(|t| t.is_numeric()) {
-                    let input_types = vec![ty.clone(); *len];
-                    // ignore error as we are just collecting all possible input types
-                    let _ = update_input_types(inner, mapping, &input_types);
-                }
-            }
-            TypeSignature::Variadic(var) => {
-                for ty in var {
-                    // assuming one or more arguments produce the same state
-                    let input_types = vec![ty.clone()];
-                    // ignore error as we are just collecting all possible input types
-                    let _ = update_input_types(inner, mapping, &input_types);
-                }
-            }
-            TypeSignature::VariadicAny => {
-                for ty in DataTypeIterator::all_flat() {
-                    // assuming one or more arguments produce the same state
-                    let input_types = vec![ty.clone()];
-                    // ignore error as we are just collecting all possible input types
-                    let _ = update_input_types(inner, mapping, &input_types);
-                }
-            }
-            TypeSignature::Uniform(cnt, tys) => {
-                for ty in tys {
-                    let input_types = vec![ty.clone(); *cnt];
-                    // ignore error as we are just collecting all possible input types
-                    let _ = update_input_types(inner, mapping, &input_types);
-                }
-            }
-            TypeSignature::UserDefined => {
-                // first determine input length
-                let input_len = match inner.name() {
-                    "sum" | "count" | "avg" | "min" | "max" => 1,
-                    _ => {
-                        return Err(datafusion_common::DataFusionError::Internal(format!(
-                            "Unsupported aggregate function for step aggr pushdown: {}",
-                            inner.name()
-                        )));
-                    }
-                };
-                match input_len {
-                    1 => {
-                        for ty in DataTypeIterator::all_flat() {
-                            let input_types = vec![ty.clone(); input_len];
-                            // ignore error as we are just collecting all possible input types
-                            let _ = update_input_types(inner, mapping, &input_types);
-                        }
-                    }
-                    _ => {
-                        return Err(datafusion_common::DataFusionError::Internal(format!(
-                            "Unsupported input length for step aggr pushdown: {}",
-                            input_len
-                        )));
-                    }
-                }
-            }
-            TypeSignature::String(cnt) => {
-                for ty in DataTypeIterator::all_flat() {
-                    let input_types = vec![ty.clone(); *cnt];
-                    // ignore error as we are just collecting all possible input types
-                    let _ = update_input_types(inner, mapping, &input_types);
-                }
-            }
-            TypeSignature::Nullary => {
-                let input_types = vec![];
-                // ignore error as we are just collecting all possible input types
-                let _ = update_input_types(inner, mapping, &input_types);
-            }
-            _ => {
-                return Err(datafusion_common::DataFusionError::Internal(format!(
-                    "Unsupported type signature for step aggr pushdown: {:?}",
-                    sig
-                )))
-            }
-        }
-        Ok(())
-    }
-    parse_ty_sig(inner, &inner.signature().type_signature, &mut mapping)?;
-
-    Ok(StateTypeLookup { mapping })
-}
-
-pub struct DataTypeIterator {
-    inner: Vec<DataType>,
-}
-
-impl DataTypeIterator {
-    pub fn new(inner: Vec<DataType>) -> Self {
-        Self { inner }
-    }
-
-    pub fn all_flat() -> Self {
-        use arrow::datatypes::{IntervalUnit, TimeUnit};
-        use DataType::*;
-        let inner = vec![
-            Null,
-            Boolean,
-            Int8,
-            Int16,
-            Int32,
-            Int64,
-            UInt8,
-            UInt16,
-            UInt32,
-            UInt64,
-            Float32,
-            Float64,
-            Decimal128(38, 18),
-            Decimal256(76, 38),
-            Binary,
-            Date32,
-            Date64,
-            Timestamp(TimeUnit::Second, None),
-            Timestamp(TimeUnit::Millisecond, None),
-            Timestamp(TimeUnit::Microsecond, None),
-            Timestamp(TimeUnit::Nanosecond, None),
-            Time32(TimeUnit::Second),
-            Time32(TimeUnit::Millisecond),
-            Time64(TimeUnit::Microsecond),
-            Time64(TimeUnit::Nanosecond),
-            Duration(TimeUnit::Second),
-            Duration(TimeUnit::Millisecond),
-            Duration(TimeUnit::Microsecond),
-            Duration(TimeUnit::Nanosecond),
-            Interval(IntervalUnit::DayTime),
-            Interval(IntervalUnit::YearMonth),
-            Interval(IntervalUnit::MonthDayNano),
-            LargeBinary,
-            BinaryView,
-            Utf8,
-            LargeUtf8,
-            Utf8View,
-        ];
-        Self { inner }
-    }
-}
-
-impl Iterator for DataTypeIterator {
-    type Item = DataType;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.inner.pop()
-    }
-}
-
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct MergeWrapper {
     inner: AggregateUDF,
@@ -727,600 +449,4 @@ impl Accumulator for MergeAccum {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::any::Any;
-    use std::pin::Pin;
-    use std::sync::Arc;
-    use std::task::{Context, Poll};
-
-    use arrow::array::{Float64Array, Int64Array, UInt64Array};
-    use arrow::record_batch::RecordBatch;
-    use arrow_schema::SchemaRef;
-    use datafusion::catalog::{Session, TableProvider};
-    use datafusion::datasource::DefaultTableSource;
-    use datafusion::execution::{RecordBatchStream, SendableRecordBatchStream, TaskContext};
-    use datafusion::optimizer::analyzer::type_coercion::TypeCoercion;
-    use datafusion::optimizer::AnalyzerRule;
-    use datafusion::physical_plan::aggregates::AggregateExec;
-    use datafusion::physical_plan::execution_plan::{Boundedness, EmissionType};
-    use datafusion::physical_plan::{DisplayAs, DisplayFormatType, ExecutionPlan, PlanProperties};
-    use datafusion::physical_planner::{DefaultPhysicalPlanner, PhysicalPlanner};
-    use datafusion::prelude::SessionContext;
-    use datafusion_common::{Column, DFSchema, TableReference};
-    use datafusion_expr::expr::AggregateFunction;
-    use datafusion_expr::{Aggregate, Expr, LogicalPlan, Signature, TableScan};
-    use datafusion_physical_expr::{EquivalenceProperties, Partitioning};
-    use futures::Stream;
-    use pretty_assertions::assert_eq;
-
-    use super::*;
-
-    #[derive(Debug)]
-    pub struct MockInputExec {
-        input: Vec<RecordBatch>,
-        schema: SchemaRef,
-        properties: PlanProperties,
-    }
-
-    impl MockInputExec {
-        pub fn new(input: Vec<RecordBatch>, schema: SchemaRef) -> Self {
-            Self {
-                properties: PlanProperties::new(
-                    EquivalenceProperties::new(schema.clone()),
-                    Partitioning::UnknownPartitioning(1),
-                    EmissionType::Incremental,
-                    Boundedness::Bounded,
-                ),
-                input,
-                schema,
-            }
-        }
-    }
-
-    impl DisplayAs for MockInputExec {
-        fn fmt_as(&self, _t: DisplayFormatType, _f: &mut std::fmt::Formatter) -> std::fmt::Result {
-            unimplemented!()
-        }
-    }
-
-    impl ExecutionPlan for MockInputExec {
-        fn name(&self) -> &str {
-            "MockInputExec"
-        }
-
-        fn as_any(&self) -> &dyn Any {
-            self
-        }
-
-        fn properties(&self) -> &PlanProperties {
-            &self.properties
-        }
-
-        fn children(&self) -> Vec<&Arc<dyn ExecutionPlan>> {
-            vec![]
-        }
-
-        fn with_new_children(
-            self: Arc<Self>,
-            _children: Vec<Arc<dyn ExecutionPlan>>,
-        ) -> datafusion_common::Result<Arc<dyn ExecutionPlan>> {
-            Ok(self)
-        }
-
-        fn execute(
-            &self,
-            _partition: usize,
-            _context: Arc<TaskContext>,
-        ) -> datafusion_common::Result<SendableRecordBatchStream> {
-            let stream = MockStream {
-                stream: self.input.clone(),
-                schema: self.schema.clone(),
-                idx: 0,
-            };
-            Ok(Box::pin(stream))
-        }
-    }
-
-    struct MockStream {
-        stream: Vec<RecordBatch>,
-        schema: SchemaRef,
-        idx: usize,
-    }
-
-    impl Stream for MockStream {
-        type Item = datafusion_common::Result<RecordBatch>;
-        fn poll_next(
-            mut self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<Option<datafusion_common::Result<RecordBatch>>> {
-            if self.idx < self.stream.len() {
-                let ret = self.stream[self.idx].clone();
-                self.idx += 1;
-                Poll::Ready(Some(Ok(ret)))
-            } else {
-                Poll::Ready(None)
-            }
-        }
-    }
-
-    impl RecordBatchStream for MockStream {
-        fn schema(&self) -> SchemaRef {
-            self.schema.clone()
-        }
-    }
-
-    #[derive(Debug)]
-    struct DummyTableProvider {
-        schema: Arc<arrow_schema::Schema>,
-    }
-
-    impl DummyTableProvider {
-        pub fn new(schema: Arc<arrow_schema::Schema>) -> Self {
-            Self { schema }
-        }
-    }
-
-    impl Default for DummyTableProvider {
-        fn default() -> Self {
-            Self {
-                schema: Arc::new(arrow_schema::Schema::new(vec![Field::new(
-                    "number",
-                    DataType::Int64,
-                    true,
-                )])),
-            }
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl TableProvider for DummyTableProvider {
-        fn as_any(&self) -> &dyn std::any::Any {
-            self
-        }
-
-        fn schema(&self) -> Arc<arrow_schema::Schema> {
-            self.schema.clone()
-        }
-
-        fn table_type(&self) -> datafusion_expr::TableType {
-            datafusion_expr::TableType::Base
-        }
-
-        async fn scan(
-            &self,
-            _state: &dyn Session,
-            _projection: Option<&Vec<usize>>,
-            _filters: &[Expr],
-            _limit: Option<usize>,
-        ) -> datafusion::error::Result<Arc<dyn ExecutionPlan>> {
-            Ok(Arc::new(MockInputExec::new(vec![], self.schema.clone())))
-        }
-    }
-
-    fn dummy_table_scan() -> LogicalPlan {
-        let table_provider = Arc::new(DummyTableProvider::default());
-        let table_source = DefaultTableSource::new(table_provider);
-        LogicalPlan::TableScan(
-            TableScan::try_new(
-                TableReference::bare("Number"),
-                Arc::new(table_source),
-                None,
-                vec![],
-                None,
-            )
-            .unwrap(),
-        )
-    }
-
-    fn dummy_table_scan_with_schema(
-        schema: Arc<DFSchema>,
-        table_name: impl Into<TableReference>,
-    ) -> LogicalPlan {
-        let table_provider = Arc::new(DummyTableProvider::new(schema.inner().clone()));
-        let table_source = DefaultTableSource::new(table_provider);
-        LogicalPlan::TableScan(
-            TableScan::try_new(table_name, Arc::new(table_source), None, vec![], None).unwrap(),
-        )
-    }
-
-    fn dummy_aggr_state_plan(state_udf: Arc<AggregateUDF>) -> LogicalPlan {
-        let table_scan = dummy_table_scan();
-        let aggr_func = AggregateFunction::new_udf(
-            state_udf,
-            vec![Expr::Column(Column::new_unqualified("number"))],
-            false,
-            None,
-            None,
-            None,
-        );
-        LogicalPlan::Aggregate(
-            Aggregate::try_new(
-                Arc::new(table_scan),
-                vec![],
-                vec![Expr::AggregateFunction(aggr_func)],
-            )
-            .unwrap(),
-        )
-    }
-
-    fn dummy_aggr_merge_plan(input_schema: Arc<DFSchema>, func: Arc<AggregateUDF>) -> LogicalPlan {
-        let table = DummyTableProvider::new(input_schema.inner().clone());
-        let table_source = DefaultTableSource::new(Arc::new(table));
-        let table_scan = LogicalPlan::TableScan(
-            TableScan::try_new(
-                TableReference::bare("MergeState"),
-                Arc::new(table_source),
-                None,
-                vec![],
-                None,
-            )
-            .unwrap(),
-        );
-        let col_name = input_schema.field(0).name().clone();
-        let column = Expr::Column(Column::new_unqualified(&col_name));
-        let aggr_func = AggregateFunction::new_udf(func, vec![column], false, None, None, None);
-        LogicalPlan::Aggregate(
-            Aggregate::try_new(
-                Arc::new(table_scan),
-                vec![],
-                vec![Expr::AggregateFunction(aggr_func)],
-            )
-            .unwrap(),
-        )
-    }
-
-    #[tokio::test]
-    async fn test_sum_udaf() {
-        let ctx = SessionContext::new();
-
-        let sum = datafusion::functions_aggregate::sum::sum_udaf();
-        let sum = (*sum).clone();
-
-        let sum_state_to_input = get_possible_state_to_input_types(&sum).unwrap();
-
-        let expected_sum_state_to_input = StateTypeLookup {
-            mapping: BTreeMap::from([
-                (vec![DataType::Float64], vec![DataType::Float64]),
-                (vec![DataType::Int64], vec![DataType::Int64]),
-                (vec![DataType::UInt64], vec![DataType::UInt64]),
-                (
-                    vec![DataType::Decimal128(38, 18)],
-                    vec![DataType::Decimal128(38, 18)],
-                ),
-                (
-                    vec![DataType::Decimal256(76, 38)],
-                    vec![DataType::Decimal256(76, 38)],
-                ),
-            ]),
-        };
-
-        assert_eq!(sum_state_to_input, expected_sum_state_to_input);
-
-        let wrapper = StateMergeWrapper::new(sum.clone(), sum_state_to_input.clone()).unwrap();
-        let expected = StateMergeWrapper {
-            original: sum.clone(),
-            state_function: StateWrapper {
-                inner: sum.clone(),
-                name: "__sum_state".to_string(),
-            },
-            merge_function: MergeWrapper {
-                inner: sum.clone(),
-                name: "__sum_merge".to_string(),
-                merge_signature: Signature::user_defined(datafusion_expr::Volatility::Immutable),
-                state_to_input_types: sum_state_to_input.clone(),
-            },
-        };
-        assert_eq!(wrapper, expected);
-
-        // create the child logical plan for the state function, which usually should be running on datanode
-        let aggr_state_plan =
-            dummy_aggr_state_plan(Arc::new(wrapper.state_function().clone().into()));
-        let expected_plan = LogicalPlan::Aggregate(
-            Aggregate::try_new(
-                Arc::new(dummy_table_scan()),
-                vec![],
-                vec![Expr::AggregateFunction(AggregateFunction::new_udf(
-                    Arc::new(wrapper.state_function().clone().into()),
-                    vec![Expr::Column(Column::new_unqualified("number"))],
-                    false,
-                    None,
-                    None,
-                    None,
-                ))],
-            )
-            .unwrap(),
-        );
-        assert_eq!(aggr_state_plan, expected_plan);
-        assert_eq!(
-            aggr_state_plan.schema().as_arrow(),
-            &arrow_schema::Schema::new(vec![Field::new(
-                "__sum_state(number)",
-                DataType::Struct(vec![Field::new("sum[sum]", DataType::Int64, true)].into()),
-                true,
-            )])
-        );
-
-        // the state schema should be the input schema of the merge function
-        let state_schema = aggr_state_plan.schema().clone();
-
-        let expected_merge_plan = LogicalPlan::Aggregate(
-            Aggregate::try_new(
-                Arc::new(dummy_table_scan_with_schema(
-                    state_schema.clone(),
-                    TableReference::bare("MergeState"),
-                )),
-                vec![],
-                vec![Expr::AggregateFunction(AggregateFunction::new_udf(
-                    Arc::new(wrapper.merge_function().clone().into()),
-                    vec![Expr::Column(Column::new_unqualified("__sum_state(number)"))],
-                    false,
-                    None,
-                    None,
-                    None,
-                ))],
-            )
-            .unwrap(),
-        );
-
-        let aggr_merge_plan = dummy_aggr_merge_plan(
-            state_schema,
-            Arc::new(wrapper.merge_function().clone().into()),
-        );
-        assert_eq!(aggr_merge_plan, expected_merge_plan);
-
-        // evaluate the state function
-        let input = Int64Array::from(vec![Some(1), Some(2), None, Some(3)]);
-        let values = vec![Arc::new(input) as arrow::array::ArrayRef];
-
-        let phy_aggr_state_plan = DefaultPhysicalPlanner::default()
-            .create_physical_plan(&aggr_state_plan, &ctx.state())
-            .await
-            .unwrap();
-        let aggr_exec = phy_aggr_state_plan
-            .as_any()
-            .downcast_ref::<AggregateExec>()
-            .unwrap();
-        let aggr_func_expr = &aggr_exec.aggr_expr()[0];
-        let mut state_accum = aggr_func_expr.create_accumulator().unwrap();
-
-        state_accum.update_batch(&values).unwrap();
-        let state = state_accum.state().unwrap();
-        assert_eq!(state.len(), 1);
-        assert_eq!(state[0], ScalarValue::Int64(Some(6)));
-
-        let eval_res = state_accum.evaluate().unwrap();
-        assert_eq!(
-            eval_res,
-            ScalarValue::Struct(Arc::new(
-                StructArray::try_new(
-                    vec![Field::new("col_0", DataType::Int64, true)].into(),
-                    vec![Arc::new(Int64Array::from(vec![Some(6)]))],
-                    None,
-                )
-                .unwrap(),
-            ))
-        );
-
-        let merge_input = vec![
-            Arc::new(Int64Array::from(vec![Some(6), Some(42), None])) as arrow::array::ArrayRef
-        ];
-        let merge_input_struct_arr = StructArray::try_new(
-            vec![Field::new("state[0]", DataType::Int64, true)].into(),
-            merge_input,
-            None,
-        )
-        .unwrap();
-
-        let phy_aggr_merge_plan = DefaultPhysicalPlanner::default()
-            .create_physical_plan(&aggr_merge_plan, &ctx.state())
-            .await
-            .unwrap();
-        let aggr_exec = phy_aggr_merge_plan
-            .as_any()
-            .downcast_ref::<AggregateExec>()
-            .unwrap();
-        let aggr_func_expr = &aggr_exec.aggr_expr()[0];
-        let mut merge_accum = aggr_func_expr.create_accumulator().unwrap();
-
-        merge_accum
-            .update_batch(&[Arc::new(merge_input_struct_arr)])
-            .unwrap();
-        let merge_state = merge_accum.state().unwrap();
-        assert_eq!(merge_state.len(), 1);
-        assert_eq!(merge_state[0], ScalarValue::Int64(Some(48)));
-
-        let merge_eval_res = merge_accum.evaluate().unwrap();
-        assert_eq!(merge_eval_res, ScalarValue::Int64(Some(48)));
-    }
-
-    #[tokio::test]
-    async fn test_avg_udaf() {
-        let ctx = SessionContext::new();
-
-        let avg = datafusion::functions_aggregate::average::avg_udaf();
-        let avg = (*avg).clone();
-
-        let avg_type = get_possible_state_to_input_types(&avg).unwrap();
-        let expected_avg_type = StateTypeLookup {
-            mapping: {
-                use DataType::*;
-                BTreeMap::from([
-                    (vec![UInt64, Int8], vec![Int8]),
-                    (vec![UInt64, Int16], vec![Int16]),
-                    (vec![UInt64, Int32], vec![Int32]),
-                    (vec![UInt64, Int64], vec![Int64]),
-                    (vec![UInt64, UInt8], vec![UInt8]),
-                    (vec![UInt64, UInt16], vec![UInt16]),
-                    (vec![UInt64, UInt32], vec![UInt32]),
-                    (vec![UInt64, UInt64], vec![UInt64]),
-                    (vec![UInt64, Float32], vec![Float32]),
-                    (vec![UInt64, Float64], vec![Float64]),
-                    (vec![UInt64, Decimal128(38, 18)], vec![Decimal128(38, 18)]),
-                    (vec![UInt64, Decimal256(76, 38)], vec![Decimal256(76, 38)]),
-                ])
-            },
-        };
-        assert_eq!(avg_type, expected_avg_type);
-
-        let wrapper = StateMergeWrapper::new(avg.clone(), avg_type.clone()).unwrap();
-
-        let expected = StateMergeWrapper {
-            original: avg.clone(),
-            state_function: StateWrapper {
-                inner: avg.clone(),
-                name: "__avg_state".to_string(),
-            },
-            merge_function: MergeWrapper {
-                inner: avg.clone(),
-                name: "__avg_merge".to_string(),
-                merge_signature: Signature::user_defined(datafusion_expr::Volatility::Immutable),
-                state_to_input_types: avg_type.clone(),
-            },
-        };
-        assert_eq!(wrapper, expected);
-
-        // create the child logical plan for the state function, which usually should be running on datanode
-        let aggr_state_plan =
-            dummy_aggr_state_plan(Arc::new(wrapper.state_function().clone().into()));
-        let expected_plan = LogicalPlan::Aggregate(
-            Aggregate::try_new(
-                Arc::new(dummy_table_scan()),
-                vec![],
-                vec![Expr::AggregateFunction(AggregateFunction::new_udf(
-                    Arc::new(wrapper.state_function().clone().into()),
-                    vec![Expr::Column(Column::new_unqualified("number"))],
-                    false,
-                    None,
-                    None,
-                    None,
-                ))],
-            )
-            .unwrap(),
-        );
-        assert_eq!(aggr_state_plan, expected_plan);
-        assert_eq!(
-            aggr_state_plan.schema().as_arrow(),
-            &arrow_schema::Schema::new(vec![Field::new(
-                "__avg_state(number)",
-                DataType::Struct(
-                    vec![
-                        Field::new("avg[count]", DataType::UInt64, true),
-                        Field::new("avg[sum]", DataType::Float64, true)
-                    ]
-                    .into()
-                ),
-                true,
-            )])
-        );
-
-        // the state schema should be the input schema of the merge function
-        let state_schema = aggr_state_plan.schema().clone();
-
-        let expected_merge_plan = LogicalPlan::Aggregate(
-            Aggregate::try_new(
-                Arc::new(dummy_table_scan_with_schema(
-                    state_schema.clone(),
-                    TableReference::bare("MergeState"),
-                )),
-                vec![],
-                vec![Expr::AggregateFunction(AggregateFunction::new_udf(
-                    Arc::new(wrapper.merge_function().clone().into()),
-                    vec![Expr::Column(Column::new_unqualified("__avg_state(number)"))],
-                    false,
-                    None,
-                    None,
-                    None,
-                ))],
-            )
-            .unwrap(),
-        );
-
-        let aggr_merge_plan = dummy_aggr_merge_plan(
-            state_schema,
-            Arc::new(wrapper.merge_function().clone().into()),
-        );
-        assert_eq!(aggr_merge_plan, expected_merge_plan);
-
-        // evaluate the state function
-        let input = Float64Array::from(vec![Some(1.), Some(2.), None, Some(3.)]);
-        let values = vec![Arc::new(input) as arrow::array::ArrayRef];
-
-        // type coerced so avg aggr function can function correctly
-        let coerced_aggr_state_plan = TypeCoercion::new()
-            .analyze(aggr_state_plan, &Default::default())
-            .unwrap();
-
-        let phy_aggr_state_plan = DefaultPhysicalPlanner::default()
-            .create_physical_plan(&coerced_aggr_state_plan, &ctx.state())
-            .await
-            .unwrap();
-        let aggr_exec = phy_aggr_state_plan
-            .as_any()
-            .downcast_ref::<AggregateExec>()
-            .unwrap();
-        let aggr_func_expr = &aggr_exec.aggr_expr()[0];
-        let mut state_accum = aggr_func_expr.create_accumulator().unwrap();
-
-        state_accum.update_batch(&values).unwrap();
-        let state = state_accum.state().unwrap();
-        assert_eq!(state.len(), 2);
-        assert_eq!(state[0], ScalarValue::UInt64(Some(3)));
-        assert_eq!(state[1], ScalarValue::Float64(Some(6.)));
-
-        let eval_res = state_accum.evaluate().unwrap();
-        let expected = Arc::new(
-            StructArray::try_new(
-                vec![
-                    Field::new("col_0", DataType::UInt64, true),
-                    Field::new("col_1", DataType::Float64, true),
-                ]
-                .into(),
-                vec![
-                    Arc::new(UInt64Array::from(vec![Some(3)])),
-                    Arc::new(Float64Array::from(vec![Some(6.)])),
-                ],
-                None,
-            )
-            .unwrap(),
-        );
-        assert_eq!(eval_res, ScalarValue::Struct(expected));
-
-        let merge_input = vec![
-            Arc::new(UInt64Array::from(vec![Some(3), Some(42), None])) as arrow::array::ArrayRef,
-            Arc::new(Float64Array::from(vec![Some(48.), Some(84.), None])),
-        ];
-        let merge_input_struct_arr = StructArray::try_new(
-            vec![
-                Field::new("state[0]", DataType::UInt64, true),
-                Field::new("state[1]", DataType::Float64, true),
-            ]
-            .into(),
-            merge_input,
-            None,
-        )
-        .unwrap();
-
-        let phy_aggr_merge_plan = DefaultPhysicalPlanner::default()
-            .create_physical_plan(&aggr_merge_plan, &ctx.state())
-            .await
-            .unwrap();
-        let aggr_exec = phy_aggr_merge_plan
-            .as_any()
-            .downcast_ref::<AggregateExec>()
-            .unwrap();
-        let aggr_func_expr = &aggr_exec.aggr_expr()[0];
-        let mut merge_accum = aggr_func_expr.create_accumulator().unwrap();
-
-        merge_accum
-            .update_batch(&[Arc::new(merge_input_struct_arr)])
-            .unwrap();
-        let merge_state = merge_accum.state().unwrap();
-        assert_eq!(merge_state.len(), 2);
-        assert_eq!(merge_state[0], ScalarValue::UInt64(Some(45)));
-        assert_eq!(merge_state[1], ScalarValue::Float64(Some(132.)));
-
-        let merge_eval_res = merge_accum.evaluate().unwrap();
-        // the merge function returns the average, which is 132 / 45
-        assert_eq!(merge_eval_res, ScalarValue::Float64(Some(132. / 45_f64)));
-    }
-}
+mod tests;
